Move page-numbers component to addon's folder
